refactor(cli): split desired-state validation
