+++ crates/gents-cli/src/desired_state/tests.rs
@@ -533,10 +533,29 @@ fn sample_event_trigger() -> DesiredEventTrigger {
     }
 }
 
-fn empty_manifest_with_event_trigger(t: DesiredEventTrigger) -> DesiredStateManifest {
-    let mut m = empty_manifest("did:test:test");
-    m.event_triggers.push(t);
-    m
+fn sample_event_trigger_for(trigger_id: &str, task_id: &str) -> DesiredEventTrigger {
+    DesiredEventTrigger {
+        trigger_id: trigger_id.to_string(),
+        task_id: task_id.to_string(),
+        source_collection: "CustomerSignup".to_string(),
+        event_kind: "created".to_string(),
+        filter: None,
+        correlation_field: None,
+        fire_mode: None,
+        expected_count: None,
+        expected_count_field: None,
+        group_timeout_secs: None,
+        group_min_count: None,
+        workspace_authority: None,
+        enabled: true,
+        concurrency: "serial".to_string(),
+    }
+}
+
+fn empty_manifest_with_event_trigger(trigger: DesiredEventTrigger) -> DesiredStateManifest {
+    let mut manifest = empty_manifest("did:test:test");
+    manifest.event_triggers.push(trigger);
+    manifest
 }
 
 #[test]
@@ -1630,1112 +1649,8 @@ fn validation_errors(manifest: &DesiredStateManifest) -> Vec<String> {
     errors
 }
 
-#[test]
-fn validate_rejects_non_positive_stream_liveness_timeout() {
-    let mut manifest = empty_manifest("did:test:test");
-    let mut profile = profile("fast");
-    profile.stream_liveness_timeout_secs = Some(0);
-    manifest.inference_profiles.push(profile);
-
-    let errors = validation_errors(&manifest);
-
-    assert!(
-        errors
-            .iter()
-            .any(|message| message.contains("stream_liveness_timeout_secs must be positive")),
-        "expected stream_liveness_timeout_secs validation error, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_zero_deadline_without_relationship_error() {
-    let mut manifest = empty_manifest("did:test:test");
-    let mut profile = profile("zero-deadline");
-    profile.stream_liveness_timeout_secs = Some(300);
-    profile.deadline_duration_secs = Some(0);
-    manifest.inference_profiles.push(profile);
-
-    let errors = validation_errors(&manifest);
-
-    assert!(
-        errors
-            .iter()
-            .any(|message| message.contains("deadline_duration_secs must be positive")),
-        "expected deadline validation error, got {errors:?}"
-    );
-    assert!(
-        !errors
-            .iter()
-            .any(|message| message.contains("must be less than deadline_duration_secs")),
-        "expected no relationship error for an invalid deadline, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_negative_deadline_without_relationship_error() {
-    let mut manifest = empty_manifest("did:test:test");
-    let mut profile = profile("negative-deadline");
-    profile.stream_liveness_timeout_secs = Some(300);
-    profile.deadline_duration_secs = Some(-1);
-    manifest.inference_profiles.push(profile);
-
-    let errors = validation_errors(&manifest);
-
-    assert!(
-        errors
-            .iter()
-            .any(|message| message.contains("deadline_duration_secs must be positive")),
-        "expected deadline validation error, got {errors:?}"
-    );
-    assert!(
-        !errors
-            .iter()
-            .any(|message| message.contains("must be less than deadline_duration_secs")),
-        "expected no relationship error for an invalid deadline, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_reports_each_invalid_timeout_without_relationship_error() {
-    let mut manifest = empty_manifest("did:test:test");
-    let mut profile = profile("invalid-timeouts");
-    profile.stream_liveness_timeout_secs = Some(0);
-    profile.deadline_duration_secs = Some(0);
-    manifest.inference_profiles.push(profile);
-
-    let errors = validation_errors(&manifest);
-
-    assert!(
-        errors
-            .iter()
-            .any(|message| message.contains("stream_liveness_timeout_secs must be positive")),
-        "expected liveness validation error, got {errors:?}"
-    );
-    assert!(
-        errors
-            .iter()
-            .any(|message| message.contains("deadline_duration_secs must be positive")),
-        "expected deadline validation error, got {errors:?}"
-    );
-    assert!(
-        !errors
-            .iter()
-            .any(|message| message.contains("must be less than deadline_duration_secs")),
-        "expected no relationship error for invalid timeouts, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_stream_liveness_timeout_equal_to_deadline() {
-    let mut manifest = empty_manifest("did:test:test");
-    let mut profile = profile("equal-timeouts");
-    profile.stream_liveness_timeout_secs = Some(300);
-    profile.deadline_duration_secs = Some(300);
-    manifest.inference_profiles.push(profile);
-
-    let errors = validation_errors(&manifest);
-
-    assert!(
-        errors.iter().any(|message| {
-            message.contains(
-                "stream_liveness_timeout_secs (300) must be less than deadline_duration_secs (300)",
-            )
-        }),
-        "expected ineffective liveness timeout validation error, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_stream_liveness_timeout_greater_than_deadline() {
-    let mut manifest = empty_manifest("did:test:test");
-    let mut profile = profile("late-liveness");
-    profile.stream_liveness_timeout_secs = Some(600);
-    profile.deadline_duration_secs = Some(300);
-    manifest.inference_profiles.push(profile);
-
-    let errors = validation_errors(&manifest);
-
-    assert!(
-        errors.iter().any(|message| {
-            message.contains(
-                "stream_liveness_timeout_secs (600) must be less than deadline_duration_secs (300)",
-            )
-        }),
-        "expected late liveness timeout validation error, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_accepts_stream_liveness_timeout_shorter_than_deadline() {
-    let mut manifest = empty_manifest("did:test:test");
-    let mut profile = profile("early-liveness");
-    profile.stream_liveness_timeout_secs = Some(300);
-    profile.deadline_duration_secs = Some(600);
-    manifest.inference_profiles.push(profile);
-
-    let errors = validation_errors(&manifest);
-
-    assert!(
-        !errors
-            .iter()
-            .any(|message| message.contains("stream_liveness_timeout_secs")),
-        "expected valid timeout relationship, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_accepts_default_inference_timeouts() {
-    let mut manifest = empty_manifest("did:test:test");
-    manifest
-        .inference_profiles
-        .push(profile("default-timeouts"));
-
-    let errors = validation_errors(&manifest);
-
-    assert!(
-        !errors
-            .iter()
-            .any(|message| message.contains("stream_liveness_timeout_secs")),
-        "expected shipped default timeout relationship to remain valid, got {errors:?}"
-    );
-}
-
-#[test]
-fn shipped_long_running_demo_manifests_validate() {
-    let workspace_root = PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("../..");
-
-    for relative_root in [
-        "demo/security-scan",
-        "demo/defending-code",
-        "demo/repo-maintenance",
-    ] {
-        let root = workspace_root.join(relative_root);
-        let (_, report) = load::load_manifest_root(&root);
-        assert!(
-            report.ok,
-            "expected {relative_root} to validate, got {:?}",
-            report.errors
-        );
-    }
-}
-
-#[test]
-fn validate_rejects_default_liveness_when_explicit_deadline_is_too_short() {
-    let mut manifest = empty_manifest("did:test:test");
-    let mut profile = profile("implicit-liveness");
-    profile.deadline_duration_secs = Some(gents::DEFAULT_STREAM_LIVENESS_TIMEOUT_SECS as i64);
-    manifest.inference_profiles.push(profile);
-
-    let errors = validation_errors(&manifest);
-
-    assert!(
-        errors.iter().any(|message| {
-            message.contains("stream_liveness_timeout_secs (1800)")
-                && message.contains("deadline_duration_secs (1800)")
-        }),
-        "expected default liveness relationship validation error, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_negative_sampling_seed() {
-    let mut manifest = empty_manifest("did:test:test");
-    let mut profile = profile("seeded");
-    profile.seed = Some(-1);
-    manifest.inference_profiles.push(profile);
-
-    let errors = validation_errors(&manifest);
-
-    assert!(
-        errors
-            .iter()
-            .any(|message| message.contains("seed must be non-negative")),
-        "expected seed validation error, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_empty_string_in_subagent_targets() {
-    let mut manifest = manifest_with_default_behavior();
-    let mut sel = sample_tool_selection("agent-tools");
-    sel.agent_did = "did:test:test".to_string();
-    sel.subagent_targets = vec!["".to_string()];
-    manifest.tool_selections.push(sel);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors
-            .iter()
-            .any(|msg| msg.contains("subagent_targets") && msg.contains("agent-tools")),
-        "expected empty subagent_targets entry rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_subagent_spawn_enabled_without_targets() {
-    let mut manifest = manifest_with_default_behavior();
-    let mut sel = sample_tool_selection("agent-tools");
-    sel.agent_did = "did:test:test".to_string();
-    sel.subagent_spawn_enabled = true;
-    sel.subagent_targets = Vec::new();
-    manifest.tool_selections.push(sel);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|msg| {
-            msg.contains("agent-tools")
-                && msg.contains("subagent_spawn_enabled")
-                && msg.contains("subagent_targets")
-        }),
-        "expected subagent_spawn_enabled-without-targets rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_subagent_spawn_enabled_with_empty_targets_vec() {
-    let mut manifest = manifest_with_default_behavior();
-    let mut sel = sample_tool_selection("agent-tools");
-    sel.agent_did = "did:test:test".to_string();
-    sel.subagent_spawn_enabled = true;
-    sel.subagent_targets = Vec::new();
-    manifest.tool_selections.push(sel);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|msg| {
-            msg.contains("agent-tools")
-                && msg.contains("subagent_spawn_enabled")
-                && msg.contains("subagent_targets")
-        }),
-        "expected subagent_spawn_enabled-with-empty-targets-vec rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_accepts_subagent_spawn_enabled_with_targets() {
-    let mut manifest = manifest_with_default_behavior();
-    let mut sel = sample_tool_selection("agent-tools");
-    sel.agent_did = "did:test:test".to_string();
-    sel.subagent_spawn_enabled = true;
-    sel.subagent_targets = vec![gents::subagent_target_entry(
-        "researcher",
-        "did:test:test",
-        "amy-research",
-        None,
-    )];
-    manifest.tool_selections.push(sel);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        !errors
-            .iter()
-            .any(|msg| msg.contains("subagent_targets") || msg.contains("subagent_spawn_enabled")),
-        "expected no subagent rejections for valid config, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_duplicate_subagent_target_name() {
-    let mut manifest = manifest_with_default_behavior();
-    let mut sel = sample_tool_selection("agent-tools");
-    sel.agent_did = "did:test:test".to_string();
-    sel.subagent_spawn_enabled = true;
-    sel.subagent_targets = vec![
-        gents::subagent_target_entry("dup", "did:test:test", "amy-research", None),
-        gents::subagent_target_entry("dup", "did:test:test", "amy-code", None),
-    ];
-    manifest.tool_selections.push(sel);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors
-            .iter()
-            .any(|msg| msg.contains("duplicate subagent target name") && msg.contains("dup")),
-        "expected duplicate-name rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_remote_did_target_when_cross_deployment_off() {
-    let mut manifest = manifest_with_default_behavior();
-    let mut sel = sample_tool_selection("agent-tools");
-    sel.agent_did = "did:test:test".to_string();
-    sel.subagent_spawn_enabled = true;
-    sel.subagent_allow_cross_deployment = false;
-    sel.subagent_targets = vec![gents::subagent_target_entry(
-        "remote-researcher",
-        "did:test:OTHER-deployment",
-        "amy-research",
-        None,
-    )];
-    manifest.tool_selections.push(sel);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|msg| {
-            msg.contains("cross-deployment subagent delegation is deferred")
-                && msg.contains("remote-researcher")
-                && msg.contains("subagent_allow_cross_deployment=true")
-        }),
-        "expected remote-DID rejection when flag is off, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_accepts_remote_did_target_when_cross_deployment_on() {
-    let mut manifest = manifest_with_default_behavior();
-    let mut sel = sample_tool_selection("agent-tools");
-    sel.agent_did = "did:test:test".to_string();
-    sel.subagent_spawn_enabled = true;
-    sel.subagent_allow_cross_deployment = true;
-    sel.subagent_targets = vec![gents::subagent_target_entry(
-        "remote-researcher",
-        "did:test:OTHER-deployment",
-        "amy-research",
-        None,
-    )];
-    manifest.tool_selections.push(sel);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        !errors
-            .iter()
-            .any(|msg| msg.contains("cross-deployment subagent delegation is deferred")),
-        "expected no cross-deployment rejection when flag is on, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_accepts_local_did_target_when_cross_deployment_off() {
-    let mut manifest = manifest_with_default_behavior();
-    let mut sel = sample_tool_selection("agent-tools");
-    sel.agent_did = "did:test:test".to_string();
-    sel.subagent_spawn_enabled = true;
-    sel.subagent_allow_cross_deployment = false;
-    sel.subagent_targets = vec![gents::subagent_target_entry(
-        "local-researcher",
-        "did:test:test",
-        "amy-research",
-        None,
-    )];
-    manifest.tool_selections.push(sel);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        !errors
-            .iter()
-            .any(|msg| msg.contains("cross-deployment subagent delegation is deferred")),
-        "expected no cross-deployment rejection for local target, got {errors:?}"
-    );
-}
-
-#[test]
-fn write_tools_deserializer_converges_object_and_string_shapes() {
-    fn deser_write_tools(write_tools: serde_json::Value) -> Vec<String> {
-        let selection = json!({
-            "selection_id": "conv-sel",
-            "agent_did": "did:test:test",
-            "enable_file_tools": false,
-            "file_tools_mode": "ReadOnly",
-            "enable_bash": false,
-            "bash_mode": "ReadOnly",
-            "enable_meta_tools": false,
-            "write_tools": write_tools,
-        });
-        let parsed: DesiredToolSelection =
-            serde_json::from_value(selection).expect("DesiredToolSelection deserializes");
-        parsed.write_tools
-    }
-
-    let object_list = deser_write_tools(json!([
-        {
-            "tool_name": "request_action",
-            "collection": "ActionRequest",
-            "fields": [{ "name": "drift_sig", "required": true }]
-        }
-    ]));
-
-    let string_list = deser_write_tools(json!([
-        "{\"collection\":\"ActionRequest\",\"fields\":[{\"required\":true,\"name\":\"drift_sig\"}],\"tool_name\":\"request_action\"}"
-    ]));
-
-    assert_eq!(
-        object_list, string_list,
-        "object-list and string-list shapes of the same decl must canonicalize to the SAME storage Vec<String> (else false apply/diff drift)"
-    );
-}
-
-fn write_tool_storage_entry(decl: &gents::WriteToolDecl) -> String {
-    serde_json::to_string(decl).expect("WriteToolDecl serializes to JSON")
-}
-
-#[test]
-fn validate_rejects_write_tool_with_invalid_collection_identifier() {
-    use gents::WriteToolDecl;
-    for collection in ["", "ActionRequest) { _docID } mutation {"] {
-        let mut manifest = manifest_with_default_behavior();
-        let mut sel = sample_tool_selection("agent-tools");
-        sel.agent_did = "did:test:test".to_string();
-        sel.write_tools = vec![write_tool_storage_entry(&WriteToolDecl {
-            tool_name: "request_action".to_string(),
-            collection: collection.to_string(),
-            description: String::new(),
-            fields: Vec::new(),
-            output_obligation: None,
-        })];
-        manifest.tool_selections.push(sel);
-
-        let errors = validation_errors(&manifest);
-        assert!(
-            errors.iter().any(|msg| msg.contains("write_tools")
-                && msg.contains("agent-tools")
-                && msg.contains("collection")),
-            "expected invalid-collection write_tools rejection, got {errors:?}"
-        );
-    }
-}
-
-#[test]
-fn validate_rejects_write_tool_with_empty_field_name() {
-    use gents::{WriteToolDecl, WriteToolField};
-    let mut manifest = manifest_with_default_behavior();
-    let mut sel = sample_tool_selection("agent-tools");
-    sel.agent_did = "did:test:test".to_string();
-    sel.write_tools = vec![write_tool_storage_entry(&WriteToolDecl {
-        tool_name: "request_action".to_string(),
-        collection: "ActionRequest".to_string(),
-        description: String::new(),
-        output_obligation: None,
-        fields: vec![WriteToolField {
-            name: "   ".to_string(),
-            required: true,
-            fill: None,
-        }],
-    })];
-    manifest.tool_selections.push(sel);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|msg| msg.contains("write_tools")
-            && msg.contains("agent-tools")
-            && msg.contains("empty name")),
-        "expected empty-field-name write_tools rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_duplicate_write_tool_name() {
-    use gents::WriteToolDecl;
-    let mut manifest = manifest_with_default_behavior();
-    let mut sel = sample_tool_selection("agent-tools");
-    sel.agent_did = "did:test:test".to_string();
-    let decl = WriteToolDecl {
-        tool_name: "request_action".to_string(),
-        collection: "ActionRequest".to_string(),
-        description: String::new(),
-        output_obligation: None,
-        fields: Vec::new(),
-    };
-    sel.write_tools = vec![
-        write_tool_storage_entry(&decl),
-        write_tool_storage_entry(&WriteToolDecl {
-            collection: "OtherCollection".to_string(),
-            ..decl
-        }),
-    ];
-    manifest.tool_selections.push(sel);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|msg| {
-            msg.contains("duplicate write_tools tool_name") && msg.contains("request_action")
-        }),
-        "expected duplicate write_tools tool_name rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_accepts_well_formed_write_tools() {
-    use gents::{WriteToolDecl, WriteToolField};
-    let mut manifest = manifest_with_default_behavior();
-    let mut sel = sample_tool_selection("agent-tools");
-    sel.agent_did = "did:test:test".to_string();
-    sel.write_tools = vec![write_tool_storage_entry(&WriteToolDecl {
-        tool_name: "request_action".to_string(),
-        collection: "ActionRequest".to_string(),
-        description: "Request a bounded action".to_string(),
-        output_obligation: None,
-        fields: vec![
-            WriteToolField {
-                name: "title".to_string(),
-                required: true,
-                fill: None,
-            },
-            WriteToolField {
-                name: "detail".to_string(),
-                required: false,
-                fill: None,
-            },
-        ],
-    })];
-    manifest.tool_selections.push(sel);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        !errors.iter().any(|msg| msg.contains("write_tools")),
-        "expected no write_tools rejections for a well-formed decl, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_zero_write_tool_output_obligation() {
-    use gents::document_config::{WriteToolOutputObligation, WriteToolOutputObligationScope};
-    use gents::WriteToolDecl;
-    let mut manifest = manifest_with_default_behavior();
-    let mut sel = sample_tool_selection("agent-tools");
-    sel.agent_did = "did:test:test".to_string();
-    sel.write_tools = vec![write_tool_storage_entry(&WriteToolDecl {
-        tool_name: "request_action".to_string(),
-        collection: "ActionRequest".to_string(),
-        description: String::new(),
-        fields: Vec::new(),
-        output_obligation: Some(WriteToolOutputObligation {
-            scope: WriteToolOutputObligationScope::Trigger,
-            minimum_writes: 0,
-            expected_count_field: None,
-        }),
-    })];
-    manifest.tool_selections.push(sel);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors
-            .iter()
-            .any(|error| error.contains("output_obligation.minimum_writes")),
-        "{errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_write_tool_name_colliding_with_builtin() {
-    use gents::{WriteToolDecl, WriteToolField};
-    let mut manifest = manifest_with_default_behavior();
-    let mut sel = sample_tool_selection("agent-tools");
-    sel.agent_did = "did:test:test".to_string();
-    sel.write_tools = vec![write_tool_storage_entry(&WriteToolDecl {
-        tool_name: "read_file".to_string(),
-        collection: "AuditLog".to_string(),
-        description: String::new(),
-        output_obligation: None,
-        fields: vec![WriteToolField {
-            name: "path".to_string(),
-            required: true,
-            fill: None,
-        }],
-    })];
-    manifest.tool_selections.push(sel);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|msg| msg.contains("write_tools")
-            && msg.contains("read_file")
-            && msg.contains("built-in")),
-        "expected built-in collision rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_write_tool_name_colliding_with_cli_tool() {
-    use gents::WriteToolDecl;
-    let mut manifest = manifest_with_default_behavior();
-    let mut sel = sample_tool_selection("agent-tools");
-    sel.agent_did = "did:test:test".to_string();
-    sel.cli_tool_names = vec!["rg".to_string()];
-    sel.write_tools = vec![write_tool_storage_entry(&WriteToolDecl {
-        tool_name: "rg".to_string(),
-        collection: "AuditLog".to_string(),
-        description: String::new(),
-        output_obligation: None,
-        fields: Vec::new(),
-    })];
-    manifest.tool_selections.push(sel);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|msg| msg.contains("write_tools")
-            && msg.contains("rg")
-            && msg.contains("cli_tool_names")),
-        "expected cli_tool_names collision rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_duplicate_write_tool_field_name() {
-    use gents::{WriteToolDecl, WriteToolField};
-    let mut manifest = manifest_with_default_behavior();
-    let mut sel = sample_tool_selection("agent-tools");
-    sel.agent_did = "did:test:test".to_string();
-    sel.write_tools = vec![write_tool_storage_entry(&WriteToolDecl {
-        tool_name: "request_action".to_string(),
-        collection: "ActionRequest".to_string(),
-        description: String::new(),
-        output_obligation: None,
-        fields: vec![
-            WriteToolField {
-                name: "summary".to_string(),
-                required: true,
-                fill: None,
-            },
-            WriteToolField {
-                name: "summary".to_string(),
-                required: false,
-                fill: None,
-            },
-        ],
-    })];
-    manifest.tool_selections.push(sel);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|msg| msg.contains("write_tools")
-            && msg.contains("request_action")
-            && msg.contains("duplicate field name")),
-        "expected duplicate field-name rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_empty_task_id() {
-    let mut manifest = manifest_with_default_behavior();
-    let mut task = sample_task("");
-    task.task_id = String::new();
-    manifest.tasks.push(task);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors
-            .iter()
-            .any(|message| message.contains("empty task_id")),
-        "expected empty task_id rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_empty_task_behavior_id() {
-    let mut manifest = manifest_with_default_behavior();
-    let mut task = sample_task("summarize-inbox");
-    task.behavior_id = String::new();
-    manifest.tasks.push(task);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors
-            .iter()
-            .any(|message| message.contains("summarize-inbox") && message.contains("behavior_id")),
-        "expected empty behavior_id rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_duplicate_task_id() {
-    let mut manifest = manifest_with_default_behavior();
-    manifest.tasks.push(sample_task("summarize-inbox"));
-    manifest.tasks.push(sample_task("summarize-inbox"));
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|message| {
-            message.contains("duplicate task_id") && message.contains("summarize-inbox")
-        }),
-        "expected duplicate task_id rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_empty_schedule_id() {
-    let mut manifest = manifest_with_default_behavior();
-    manifest.tasks.push(sample_task("summarize-inbox"));
-    let mut schedule = sample_schedule("", "summarize-inbox");
-    schedule.schedule_id = String::new();
-    manifest.schedules.push(schedule);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors
-            .iter()
-            .any(|message| message.contains("empty schedule_id")),
-        "expected empty schedule_id rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_duplicate_schedule_id() {
-    let mut manifest = manifest_with_default_behavior();
-    manifest.tasks.push(sample_task("summarize-inbox"));
-    manifest
-        .schedules
-        .push(sample_schedule("hourly", "summarize-inbox"));
-    manifest
-        .schedules
-        .push(sample_schedule("hourly", "summarize-inbox"));
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|message| {
-            message.contains("duplicate schedule_id") && message.contains("hourly")
-        }),
-        "expected duplicate schedule_id rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_schedule_interval_zero_or_negative() {
-    let mut manifest = manifest_with_default_behavior();
-    manifest.tasks.push(sample_task("summarize-inbox"));
-    let mut schedule = sample_schedule("hourly", "summarize-inbox");
-    schedule.interval_secs = Some(0);
-    manifest.schedules.push(schedule);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors
-            .iter()
-            .any(|message| message.contains("hourly") && message.contains("interval_secs")),
-        "expected interval_secs >= 1 rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_accepts_cron_schedule_with_timezone() {
-    let mut manifest = manifest_with_default_behavior();
-    manifest.tasks.push(sample_task("summarize-inbox"));
-    let mut schedule = sample_schedule("weekday-digest", "summarize-inbox");
-    schedule.interval_secs = None;
-    schedule.cron = Some("30 3 * * MON".to_string());
-    schedule.timezone = Some("America/Los_Angeles".to_string());
-    schedule.missed_run_policy = Some("latest_only".to_string());
-    manifest.schedules.push(schedule);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.is_empty(),
-        "expected valid cron schedule, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_malformed_cron_schedule() {
-    let mut manifest = manifest_with_default_behavior();
-    manifest.tasks.push(sample_task("summarize-inbox"));
-    let mut schedule = sample_schedule("bad-cron", "summarize-inbox");
-    schedule.interval_secs = None;
-    schedule.cron = Some("30 3 * *".to_string());
-    schedule.timezone = Some("UTC".to_string());
-    manifest.schedules.push(schedule);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|message| {
-            message.contains("bad-cron")
-                && message.contains("invalid cron schedule")
-                && message.contains("exactly 5 fields")
-        }),
-        "expected malformed cron rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_invalid_cron_timezone() {
-    let mut manifest = manifest_with_default_behavior();
-    manifest.tasks.push(sample_task("summarize-inbox"));
-    let mut schedule = sample_schedule("bad-zone", "summarize-inbox");
-    schedule.interval_secs = None;
-    schedule.cron = Some("30 3 * * MON".to_string());
-    schedule.timezone = Some("Mars/Olympus".to_string());
-    manifest.schedules.push(schedule);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|message| {
-            message.contains("bad-zone")
-                && message.contains("invalid cron schedule")
-                && message.contains("invalid IANA timezone")
-        }),
-        "expected invalid timezone rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_schedule_with_interval_and_cron() {
-    let mut manifest = manifest_with_default_behavior();
-    manifest.tasks.push(sample_task("summarize-inbox"));
-    let mut schedule = sample_schedule("double-cadence", "summarize-inbox");
-    schedule.cron = Some("30 3 * * MON".to_string());
-    schedule.timezone = Some("UTC".to_string());
-    manifest.schedules.push(schedule);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|message| {
-            message.contains("double-cadence")
-                && message.contains("exactly one of interval_secs or cron")
-        }),
-        "expected double cadence rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_schedule_unknown_concurrency() {
-    let mut manifest = manifest_with_default_behavior();
-    manifest.tasks.push(sample_task("summarize-inbox"));
-    let mut schedule = sample_schedule("hourly", "summarize-inbox");
-    schedule.concurrency = "everything-everywhere".to_string();
-    manifest.schedules.push(schedule);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|message| {
-            message.contains("hourly")
-                && message.contains("concurrency")
-                && message.contains("everything-everywhere")
-        }),
-        "expected unknown concurrency rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_task_unknown_behavior() {
-    let mut manifest = manifest_with_default_behavior();
-    let mut task = sample_task("summarize-inbox");
-    task.behavior_id = "did:test:test:missing".to_string();
-    manifest.tasks.push(task);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|message| {
-            message.contains("summarize-inbox")
-                && message.contains("missing")
-                && message.contains("behavior_id")
-        }),
-        "expected missing behavior_id reference rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_schedule_task_template_referencing_doc_scope() {
-    let mut manifest = manifest_with_default_behavior();
-    let mut task = sample_task("summarize-inbox");
-    task.prompt_template = "Schedule fired at {{ event.fired_at }} for {{ doc.foo }}.".to_string();
-    manifest.tasks.push(task);
-    manifest
-        .schedules
-        .push(sample_schedule("hourly", "summarize-inbox"));
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|message| {
-            message.contains("hourly")
-                && message.contains("forbidden scope")
-                && message.contains("doc")
-                && message.contains("event.*")
-        }),
-        "expected schedule-scope rejection for doc.*, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_schedule_task_template_referencing_args_scope() {
-    let mut manifest = manifest_with_default_behavior();
-    let mut task = sample_task("summarize-inbox");
-    task.prompt_template = "{{ args.target }}".to_string();
-    manifest.tasks.push(task);
-    manifest
-        .schedules
-        .push(sample_schedule("hourly", "summarize-inbox"));
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|message| {
-            message.contains("hourly")
-                && message.contains("forbidden scope")
-                && message.contains("args")
-        }),
-        "expected schedule-scope rejection for args.*, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_accepts_schedule_task_template_using_only_event_scope() {
-    let mut manifest = manifest_with_default_behavior();
-    let mut task = sample_task("summarize-inbox");
-    task.prompt_template = "Run at {{ event.fired_at }} for {{ event.trigger_kind }}.".to_string();
-    manifest.tasks.push(task);
-    manifest
-        .schedules
-        .push(sample_schedule("hourly", "summarize-inbox"));
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        !errors
-            .iter()
-            .any(|message| message.contains("forbidden scope")),
-        "expected no schedule-scope rejections, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_schedule_unknown_task() {
-    let mut manifest = manifest_with_default_behavior();
-    manifest
-        .schedules
-        .push(sample_schedule("hourly", "missing-task"));
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|message| {
-            message.contains("hourly")
-                && message.contains("missing-task")
-                && message.contains("task_id")
-        }),
-        "expected missing task_id reference rejection, got {errors:?}"
-    );
-}
-
-fn sample_event_trigger_for(trigger_id: &str, task_id: &str) -> DesiredEventTrigger {
-    DesiredEventTrigger {
-        trigger_id: trigger_id.to_string(),
-        task_id: task_id.to_string(),
-        source_collection: "CustomerSignup".to_string(),
-        event_kind: "created".to_string(),
-        filter: None,
-        correlation_field: None,
-        fire_mode: None,
-        expected_count: None,
-        expected_count_field: None,
-        group_timeout_secs: None,
-        group_min_count: None,
-        workspace_authority: None,
-        enabled: true,
-        concurrency: "serial".to_string(),
-    }
-}
-
-#[test]
-fn validate_rejects_event_trigger_referencing_unknown_task() {
-    let mut manifest = manifest_with_default_behavior();
-    manifest.event_triggers.push(sample_event_trigger_for(
-        "new-customer-greet",
-        "missing-task",
-    ));
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|message| {
-            message.contains("new-customer-greet")
-                && message.contains("unknown task_id")
-                && message.contains("missing-task")
-        }),
-        "expected unknown task_id rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_event_trigger_unknown_event_kind() {
-    let mut manifest = manifest_with_default_behavior();
-    manifest.tasks.push(sample_task("summarize-inbox"));
-    let mut trig = sample_event_trigger_for("new-customer-greet", "summarize-inbox");
-    trig.event_kind = "updated".to_string();
-    manifest.event_triggers.push(trig);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|message| {
-            message.contains("new-customer-greet") && message.contains("unsupported event_kind")
-        }),
-        "expected unsupported event_kind rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_event_trigger_template_referencing_args_scope() {
-    let mut manifest = manifest_with_default_behavior();
-    let mut task = sample_task("summarize-inbox");
-    task.prompt_template = "{{ args.foo }}".to_string();
-    manifest.tasks.push(task);
-    manifest.event_triggers.push(sample_event_trigger_for(
-        "new-customer-greet",
-        "summarize-inbox",
-    ));
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|message| {
-            message.contains("new-customer-greet") && message.contains("forbidden scope: args")
-        }),
-        "expected event-trigger forbidden-args rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_accepts_event_trigger_template_using_event_and_doc_scopes() {
-    let mut manifest = manifest_with_default_behavior();
-    let mut task = sample_task("summarize-inbox");
-    task.prompt_template = "{{ event.fired_at }} {{ doc.name }}".to_string();
-    manifest.tasks.push(task);
-    manifest.event_triggers.push(sample_event_trigger_for(
-        "new-customer-greet",
-        "summarize-inbox",
-    ));
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        !errors
-            .iter()
-            .any(|message| message.contains("forbidden scope")),
-        "expected no forbidden-scope rejections for event+doc scopes, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_duplicate_event_trigger_id() {
-    let mut manifest = manifest_with_default_behavior();
-    manifest.tasks.push(sample_task("summarize-inbox"));
-    manifest.event_triggers.push(sample_event_trigger_for(
-        "new-customer-greet",
-        "summarize-inbox",
-    ));
-    manifest.event_triggers.push(sample_event_trigger_for(
-        "new-customer-greet",
-        "summarize-inbox",
-    ));
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|message| {
-            message.contains("duplicate") && message.contains("new-customer-greet")
-        }),
-        "expected duplicate trigger_id rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_event_trigger_unknown_concurrency() {
-    let mut manifest = manifest_with_default_behavior();
-    manifest.tasks.push(sample_task("summarize-inbox"));
-    let mut trig = sample_event_trigger_for("new-customer-greet", "summarize-inbox");
-    trig.concurrency = "weird".to_string();
-    manifest.event_triggers.push(trig);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|message| {
-            message.contains("new-customer-greet")
-                && message.contains("unknown concurrency")
-                && message.contains("weird")
-        }),
-        "expected unknown concurrency rejection, got {errors:?}"
-    );
-}
+mod live;
+mod validation;
 
 #[test]
 fn export_bundle_round_trip_preserves_tasks_and_schedules() {
@@ -3151,9 +2066,9 @@ pub(super) mod write_manifest_root {
     #[test]
     fn none_system_prompt_omits_sidecar_and_field() {
         let tmp = tempdir().unwrap();
-        let mut m = minimal_manifest();
-        m.agent_behaviors[0].system_prompt = None;
-        write_manifest_root(tmp.path(), &m, false).unwrap();
+        let mut manifest = minimal_manifest();
+        manifest.agent_behaviors[0].system_prompt = None;
+        write_manifest_root(tmp.path(), &manifest, false).unwrap();
 
         let sidecar = tmp.path().join("agent-behaviors/default/system_prompt.md");
         assert!(!sidecar.exists());
@@ -3167,9 +2082,9 @@ pub(super) mod write_manifest_root {
     #[test]
     fn rejects_behavior_with_unsafe_id() {
         let tmp = tempdir().unwrap();
-        let mut m = minimal_manifest();
-        m.agent_behaviors[0].behavior_id = "bad/id".to_string();
-        let err = write_manifest_root(tmp.path(), &m, false).unwrap_err();
+        let mut manifest = minimal_manifest();
+        manifest.agent_behaviors[0].behavior_id = "bad/id".to_string();
+        let err = write_manifest_root(tmp.path(), &manifest, false).unwrap_err();
         assert!(err.contains("filesystem-unsafe"), "got: {err}");
     }
 
@@ -3317,158 +2232,6 @@ mod write_manifest_root_safe_id {
         let err = check_filesystem_safe_id(".hidden").unwrap_err();
         assert!(err.contains("dot-prefixed"), "got: {err}");
     }
-}
-
-fn sample_surface(surface_id: &str) -> DesiredDatastoreToolSurface {
-    use gents::WriteToolDecl;
-    DesiredDatastoreToolSurface {
-        surface_id: surface_id.to_string(),
-        agent_did: "did:test:test".to_string(),
-        display_name: Some("experiment writes".to_string()),
-        enabled: true,
-        entries: vec![write_tool_storage_entry(&WriteToolDecl {
-            tool_name: "write_experiment_finding".to_string(),
-            collection: "ExperimentFinding".to_string(),
-            description: "Record a finding".to_string(),
-            output_obligation: None,
-            fields: Vec::new(),
-        })],
-    }
-}
-
-#[test]
-fn validate_accepts_surface_query_entry() {
-    use gents::{QueryToolDecl, SurfaceToolDecl, WriteToolField, WriteToolFieldFill};
-    let mut manifest = manifest_with_default_behavior();
-    let mut surface = sample_surface("scan-reads");
-    surface.entries.push(
-        serde_json::to_string(&SurfaceToolDecl::Query(QueryToolDecl {
-            tool_name: "query_candidate_finding".to_string(),
-            collection: "CandidateFinding".to_string(),
-            description: "Load candidates".to_string(),
-            fields: vec!["finding_id".to_string(), "title".to_string()],
-            filter_fields: vec![WriteToolField {
-                name: "run_id".to_string(),
-                required: false,
-                fill: Some(WriteToolFieldFill::Correlation),
-            }],
-        }))
-        .unwrap(),
-    );
-    manifest.datastore_tool_surfaces.push(surface);
-    let mut sel = sample_tool_selection("stage-tools");
-    sel.agent_did = "did:test:test".to_string();
-    sel.datastore_tool_surface_ids = vec!["scan-reads".to_string()];
-    manifest.tool_selections.push(sel);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.is_empty(),
-        "query surface entries should validate, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_query_tool_name_colliding_with_cli_tool() {
-    use gents::{QueryToolDecl, SurfaceToolDecl};
-    let mut manifest = manifest_with_default_behavior();
-    let mut surface = sample_surface("scan-reads");
-    surface.entries.push(
-        serde_json::to_string(&SurfaceToolDecl::Query(QueryToolDecl {
-            tool_name: "gh".to_string(),
-            collection: "CandidateFinding".to_string(),
-            description: "Load candidates".to_string(),
-            fields: vec!["finding_id".to_string()],
-            filter_fields: Vec::new(),
-        }))
-        .unwrap(),
-    );
-    manifest.datastore_tool_surfaces.push(surface);
-    let mut sel = sample_tool_selection("stage-tools");
-    sel.agent_did = "did:test:test".to_string();
-    sel.datastore_tool_surface_ids = vec!["scan-reads".to_string()];
-    sel.cli_tool_names = vec!["gh".to_string()];
-    manifest.tool_selections.push(sel);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|error| error.contains("cli_tool_names")),
-        "query/cli name collision should fail apply validation, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_accepts_surface_linked_tool_selection() {
-    let mut manifest = manifest_with_default_behavior();
-    manifest
-        .datastore_tool_surfaces
-        .push(sample_surface("experiment-writes"));
-    let mut sel = sample_tool_selection("stage-tools");
-    sel.agent_did = "did:test:test".to_string();
-    sel.datastore_tool_surface_ids = vec!["experiment-writes".to_string()];
-    manifest.tool_selections.push(sel);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.is_empty(),
-        "surface-linked selection should validate cleanly, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_missing_surface_link() {
-    let mut manifest = manifest_with_default_behavior();
-    let mut sel = sample_tool_selection("stage-tools");
-    sel.agent_did = "did:test:test".to_string();
-    sel.datastore_tool_surface_ids = vec!["does-not-exist".to_string()];
-    manifest.tool_selections.push(sel);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors
-            .iter()
-            .any(|msg| msg.contains("missing DatastoreToolSurface")
-                && msg.contains("does-not-exist")),
-        "expected missing surface rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_disabled_surface_link() {
-    let mut manifest = manifest_with_default_behavior();
-    let mut surface = sample_surface("experiment-writes");
-    surface.enabled = false;
-    manifest.datastore_tool_surfaces.push(surface);
-    let mut sel = sample_tool_selection("stage-tools");
-    sel.agent_did = "did:test:test".to_string();
-    sel.datastore_tool_surface_ids = vec!["experiment-writes".to_string()];
-    manifest.tool_selections.push(sel);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors.iter().any(|msg| msg.contains("disabled")),
-        "expected disabled surface rejection, got {errors:?}"
-    );
-}
-
-#[test]
-fn validate_rejects_foreign_agent_surface_link() {
-    let mut manifest = manifest_with_default_behavior();
-    let mut surface = sample_surface("experiment-writes");
-    surface.agent_did = "did:key:zOther".to_string();
-    manifest.datastore_tool_surfaces.push(surface);
-    let mut sel = sample_tool_selection("stage-tools");
-    sel.agent_did = "did:test:test".to_string();
-    sel.datastore_tool_surface_ids = vec!["experiment-writes".to_string()];
-    manifest.tool_selections.push(sel);
-
-    let errors = validation_errors(&manifest);
-    assert!(
-        errors
-            .iter()
-            .any(|msg| msg.contains("different agent") || msg.contains("does not match principal")),
-        "expected foreign surface rejection, got {errors:?}"
-    );
 }
 
 /// The checked-in pack must load with no environment set — `${VAR:-default}`
