+++ crates/gents-cli/src/desired_state/tests/validation/mod.rs
@@ -1,0 +1,3 @@
+mod agent;
+mod automation;
+mod tooling;
