+++ crates/gents-cli/src/desired_state/validate/storage.rs
@@ -1,0 +1,55 @@
+pub(crate) fn non_empty(value: &Option<String>) -> Option<&str> {
+    value
+        .as_deref()
+        .map(str::trim)
+        .filter(|value| !value.is_empty())
+}
+
+pub(crate) fn normalize_tool_service_string(value: Option<String>) -> String {
+    value.unwrap_or_default().trim().to_string()
+}
+
+pub(crate) fn normalize_tool_service_mcp_path(value: Option<String>) -> String {
+    use super::super::DEFAULT_TOOL_SERVICE_MCP_PATH;
+    let trimmed = value.as_deref().unwrap_or_default().trim();
+    if trimmed.is_empty() {
+        DEFAULT_TOOL_SERVICE_MCP_PATH.to_string()
+    } else if trimmed.starts_with('/') {
+        trimmed.to_string()
+    } else {
+        format!("/{trimmed}")
+    }
+}
+
+pub(in crate::desired_state) fn optional_string_from_value(
+    field: &str,
+    value: Option<&serde_json::Value>,
+) -> anyhow::Result<Option<String>> {
+    use anyhow::anyhow;
+    use serde_json::Value;
+    match value {
+        Some(Value::String(value)) => Ok(Some(value.clone())),
+        Some(Value::Null) | None => Ok(None),
+        Some(value) => Err(anyhow!(
+            "ToolServiceRegistry field {field} must be a string or null, got {value}"
+        )),
+    }
+}
+
+pub(in crate::desired_state) fn optional_i64_from_value(
+    field: &str,
+    value: Option<&serde_json::Value>,
+) -> anyhow::Result<Option<i64>> {
+    use anyhow::anyhow;
+    use serde_json::Value;
+    match value {
+        Some(Value::Number(value)) => value
+            .as_i64()
+            .map(Some)
+            .ok_or_else(|| anyhow!("ToolServiceRegistry field {field} must be an integer")),
+        Some(Value::Null) | None => Ok(None),
+        Some(value) => Err(anyhow!(
+            "ToolServiceRegistry field {field} must be an integer or null, got {value}"
+        )),
+    }
+}
